Set Track.cxx to use MPDMagneticField
As is says on the tin, modified Track.cxx to use MPDMagneticField to get
the BField when calculating the track momentum.

Bug fix for how MPDMagneticField handles a symmetric field map. The
origin offset is now subtracted first before applying symmetry. Now an
offset of zero is passed to the interpolation routine as it has already
been accounted for.